feat(b20-asset): ERC-8056 scaled multiplier (scheduled updates + conformance) (#173)
* tokenized stock splits: integrate erc-8056

* test(b20-asset): restore pending-multiplier codec coverage

Re-adds test/unit/storage/MockB20AssetSlotHelpers.t.sol, the sole caller of
MockB20AssetStorage.pendingMultiplierValue / pendingEffectiveAt. It had been
dropped, leaving those packed-slot decoders uncovered and failing the
Interface Coverage CI gate on #173.


* docs(IB20Asset): simplify MultiplierUpdateCancelled notice

Address review feedback: lead with what the event means and keep the
"emitted by" note; drop the wordier companion/standard-gap preamble.


---------

+++ test/unit/B20Asset/multiplier/scaledBalanceOf.t.sol
@@ -10,7 +10,7 @@ contract B20AssetScaledBalanceOfTest is B20AssetTest {
     /// @dev Property: empty balance => zero scaled balance regardless of the multiplier.
     function test_scaledBalanceOf_success_zeroForEmptyAccount(address account, uint256 newMultiplier) public {
         _assumeValidActor(account);
-        newMultiplier = bound(newMultiplier, 1, type(uint256).max);
+        newMultiplier = bound(newMultiplier, 1, type(uint128).max);
         _updateMultiplier(newMultiplier);
         assertEq(asset().scaledBalanceOf(account), 0, "empty account must have zero scaled balance");
     }
